refactor: user registrar
